refactor(qwen35): split TP-shard helpers into scheduler/tp.rs
Move tp_prefill_items/tp_decode_items/align_prefill_results/align_decode_results/
split_decode_artifacts (170 lines) out of scheduler.rs into scheduler/tp.rs. Pure move:
mod tp + use self::tp::*; pub(super) on the fns.

Validated on A100: cargo check + cargo fmt OK, cargo test --lib 95 passed /
0 failed (7 TP tests ignored).
scheduler.rs 2072 -> 1905.
